attacker_dashboard/backend/requirements.txt
Flask
Flask-CORS
requests
Pillow